# src/main/java/com/nokhrin/accounting/infrastructure/dto/DepositRequest.java
package com.nokhrin.accounting.infrastructure.dto;

import jakarta.validation.constraints.DecimalMin;
import jakarta.validation.constraints.NotNull;

import java.math.BigDecimal;

public record DepositRequest(
        @NotNull(message = "Amount is required")
        @DecimalMin(value = "0.01",  message = "Amount must be positive")
        BigDecimal amount
) {
}