# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: Accounting Service API
  version: 0.1.0
  description: REST API для управления банковскими счетами
  contact:
    name: Alexander Nokhrin
    email: a.v.nokhrin@yandex.ru
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local development

paths:
  /accounts:
    post:
      summary: Открытие нового счета
      description: Создание нового банковского счета с начальным балансом
      operationId: createAccount
      tags:
        - Accounts
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccountRequest'
      responses:
        '201':
          description: Счет успешно создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountResponse'
        '400':
          description: Ошибка валидации запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /accounts/{id}/deposit:
    post:
      summary: Пополнение счета
      operationId: deposit
      tags:
        - Accounts
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: Счет пополнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationResponse'
        '400':
          description: Ошибка валидации или неактивный счет
        '404':
          description: Счет не найден

components:
  schemas:
    CreateAccountRequest:
      type: object
      required:
        - initialBalance
      properties:
        initialBalance:
          type: number
          format: decimal
          minimum: 0
          description: Начальный баланс счета
          example: 1000.00

    AccountResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Уникальный идентификатор счета
        balance:
          type: number
          format: decimal
          description: Текущий баланс
        status:
          type: string
          enum: [ACTIVE, BLOCKED, CLOSED]
          description: Статус счета
        createdAt:
          type: string
          format: date-time
          description: Дата и время создания счета

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Описание ошибки
        timestamp:
          type: string
          format: date-time
          description: Время возникновения ошибки

    DepositRequest:
      type: object
      required: [amount]
      properties:
        amount:
          type: number
          format: decimal
          minimum: 0.01

    OperationResponse:
      type: object
      properties:
        operationId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        newBalance:
          type: number
        status:
          type: string